[openssl] move initialization into network_openssl.c
The code to initialize OpenSSL moves from "network_init" function to a
new function "network_init_openssl".
A fallback of this new function permits to call it when OpenSSL is not
configurated.
